requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
asyncpg
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
email-validator
cloudinary
python-jose
pydantic[email]